refactor: Split consumer.rs file into files in order to ease navigability
