+++ src/consumer/message.rs
@@ -1,0 +1,45 @@
+use crate::consumer::data::MessageData;
+use crate::message::proto::{MessageIdData, MessageMetadata};
+use crate::{DeserializeMessage, Payload};
+use std::marker::PhantomData;
+
+/// a message received by a consumer
+///
+/// it is generic over the type it can be deserialized to
+pub struct Message<T> {
+    /// origin topic of the message
+    pub topic: String,
+    /// contains the message's data and other metadata
+    pub payload: Payload,
+    /// contains the message's id and batch size data
+    pub message_id: MessageData,
+    pub(crate) _phantom: PhantomData<T>,
+}
+
+impl<T> Message<T> {
+    /// Pulsar metadata for the message
+    #[cfg_attr(feature = "telemetry", tracing::instrument(skip_all))]
+    pub fn metadata(&self) -> &MessageMetadata {
+        &self.payload.metadata
+    }
+
+    /// Get Pulsar message id for the message
+    #[cfg_attr(feature = "telemetry", tracing::instrument(skip_all))]
+    pub fn message_id(&self) -> &MessageIdData {
+        &self.message_id.id
+    }
+
+    /// Get message key (partition key)
+    #[cfg_attr(feature = "telemetry", tracing::instrument(skip_all))]
+    pub fn key(&self) -> Option<String> {
+        self.payload.metadata.partition_key.clone()
+    }
+}
+
+impl<T: DeserializeMessage> Message<T> {
+    /// directly deserialize a message
+    #[cfg_attr(feature = "telemetry", tracing::instrument(skip_all))]
+    pub fn deserialize(&self) -> T::Output {
+        T::deserialize_message(&self.payload)
+    }
+}
